+++ src/main/java/org/kainos/ea/cli/JobRequest.java
@@ -5,11 +5,20 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class JobRequest {
     private String roleName;
+    private String jobFamily;
     private String specification;
     private String capabilityName;
     private String bandName;
     private String responsibilities;
     private String sharepointLink;
+
+    public String getJobFamily() {
+        return jobFamily;
+    }
+
+    public void setJobFamily(String jobFamily) {
+        this.jobFamily = jobFamily;
+    }
 
     public String getRoleName() {
         return roleName;
@@ -65,12 +74,14 @@ public class JobRequest {
     @JsonCreator
     public JobRequest(
             @JsonProperty("roleName") String roleName,
+            @JsonProperty("roleName") String jobFamily,
             @JsonProperty("specification") String specification,
             @JsonProperty("capabilityName") String capabilityName,
             @JsonProperty("bandName") String bandName,
             @JsonProperty("responsibilities") String responsibilities,
             @JsonProperty("sharepointLink") String sharepointLink) {
                 this.roleName = roleName;
+                this.jobFamily = jobFamily;
                 this.specification = specification;
                 this.capabilityName = capabilityName;
                 this.bandName = bandName;
